requirements.txt
fastapi
uvicorn
jinja2
python-multipart
numpy
pypdf
python-dotenv
requests
scikit-learn
